convert to CRLF and so on

+++ src/getbit.cpp
@@ -424,11 +424,11 @@ do {                                                                    \
 
 #define GET_PES_TIMESTAMP(ts, b1, b2, b3, b4, b5)       \
 do {                                                    \
-    ts = (__int64) (b1 & 0x0e) << 29                    \
-                 | (b2       ) << 22                    \
-                 | (b3 & 0xfe) << 14                    \
-                 | (b4       ) <<  7                    \
-                 | (b5       ) >>  1;                   \
+    ts = (__int64) ((__int64)b1 & 0x0e) << 29                    \
+                 | ((__int64)b2       ) << 22                    \
+                 | ((__int64)b3 & 0xfe) << 14                    \
+                 | ((__int64)b4       ) <<  7                    \
+                 | ((__int64)b5       ) >>  1;                   \
 } while (0)
 
 // Transport packet data structure.
